students/K3339/Filatov_Arseny/Lr1/practice_1_3/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.12
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
alembic>=1.14.0
passlib[bcrypt]>=1.7.4
# bcrypt>=4.1 ломает passlib 1.7.4 (ValueError: password cannot be longer than 72 bytes при init)
bcrypt>=4.0.1,<4.1
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.6.0
